refactor(telemetry): split TelemetryDashboard into modular chart/ui/constants submodules
Extract 12 sub-components from monolithic TelemetryDashboard.tsx (444→13 lines):
- charts/: FunnelChart, FlowDistributionChart, HeroQueueChart, HourlyChart
- ui/: KpiCard, StatCard, SectionCard, SectionTitle, RangeTabs, EventTypeChip, EmptyState
- constants/: labels (PHASE_LABELS, TYPE_LABELS, etc.)

Fix import paths for domain types (../../domain vs ../../../domain).
